refactor: MAT-355 기본 mock 데이터, 임시 서버 데이터 타입을 별도로 추출
