+++ src/routes/teams/$teamId/players/-components/MemberTable/MemberTable.tsx
@@ -1,108 +1,34 @@
 import { ChevronRightIcon } from '@/assets/icons';
-import {
-  type Foot,
-  MemberFilters,
-  type Position,
-  type Role,
-} from '@/routes/teams/$teamId/players/-components';
+import noProfilePlayerImage from '@/assets/images/noProfilePlayer.png';
+import { MemberFilters } from '@/routes/teams/$teamId/players/-components/MemberFilters';
+import type {
+  Member,
+  Position,
+} from '@/routes/teams/$teamId/players/-temp-server-types';
 
 import * as styles from './MemberTable.css';
 
-interface Member {
-  id: number;
-  name: string;
-  number: string;
-  position: Position;
-  foot: Foot;
-  role: Role;
-  joinDate: string;
-  profileImage?: string;
-}
+const getPositionStyle = (position: Position): string => {
+  switch (position) {
+    case 'FW':
+      return styles.positionTagFW;
+    case 'MF':
+      return styles.positionTagMF;
+    case 'DF':
+      return styles.positionTagDF;
+    case 'GK':
+      return styles.positionTagGK;
+    default:
+      return styles.positionTagFW;
+  }
+};
 
 interface MemberTableProps {
-  members?: Member[];
+  members: Member[];
   onMemberMoreClick?: (member: Member) => void;
 }
 
-const defaultMembers: Member[] = [
-  {
-    id: 1,
-    name: '손흥민',
-    number: '#7',
-    position: 'FW',
-    foot: '왼발',
-    role: '일반',
-    joinDate: '2025-04-23',
-  },
-  {
-    id: 2,
-    name: '손흥민',
-    number: '#7',
-    position: 'MF',
-    foot: '오른발',
-    role: '일반',
-    joinDate: '2025-04-23',
-  },
-  {
-    id: 3,
-    name: '손흥민',
-    number: '#7',
-    position: 'DF',
-    foot: '양발',
-    role: '일반',
-    joinDate: '2025-04-23',
-  },
-  {
-    id: 4,
-    name: '손흥민',
-    number: '#7',
-    position: 'GK',
-    foot: '양발',
-    role: '일반',
-    joinDate: '2025-04-23',
-  },
-  {
-    id: 5,
-    name: '손흥민',
-    number: '#7',
-    position: 'FW',
-    foot: '양발',
-    role: '일반',
-    joinDate: '2025-04-23',
-  },
-  {
-    id: 6,
-    name: '손흥민',
-    number: '#7',
-    position: 'FW',
-    foot: '양발',
-    role: '일반',
-    joinDate: '2025-04-23',
-  },
-  {
-    id: 7,
-    name: '손흥민',
-    number: '#7',
-    position: 'FW',
-    foot: '양발',
-    role: '일반',
-    joinDate: '2025-04-23',
-  },
-  {
-    id: 8,
-    name: '손흥민',
-    number: '#7',
-    position: 'FW',
-    foot: '양발',
-    role: '일반',
-    joinDate: '2025-04-23',
-  },
-];
-
-export function MemberTable({
-  members = defaultMembers,
-  onMemberMoreClick,
-}: MemberTableProps) {
+export function MemberTable({ members, onMemberMoreClick }: MemberTableProps) {
   return (
     <div className={styles.container}>
       <MemberFilters />
@@ -131,12 +57,9 @@ export function MemberTable({
                   <div className={styles.leftSection}>
                     <div className={styles.indexNumber}>{index + 1}</div>
                     <img
-                      src={
-                        member.profileImage ||
-                        'https://cdn.builder.io/api/v1/image/assets/38cbf5816d9b4e569facb33a6b794634/e74f0710b5ab267430e810eabb2122e5690bd3ed?placeholderIfAbsent=true'
-                      }
+                      src={member.profileImage ?? noProfilePlayerImage}
                       className={styles.profileImage}
-                      alt={`${member.name} 프로필`}
+                      alt=''
                     />
                     <div className={styles.memberName}>{member.name}</div>
                     <div className={styles.memberNumber}>{member.number}</div>
@@ -167,18 +90,3 @@ export function MemberTable({
     </div>
   );
 }
-
-const getPositionStyle = (position: Position): string => {
-  switch (position) {
-    case 'FW':
-      return styles.positionTagFW;
-    case 'MF':
-      return styles.positionTagMF;
-    case 'DF':
-      return styles.positionTagDF;
-    case 'GK':
-      return styles.positionTagGK;
-    default:
-      return styles.positionTagFW;
-  }
-};
