+++ src/routes/teams/$teamId/players/-mock-data.ts
@@ -1,0 +1,84 @@
+import type { Member } from '@/routes/teams/$teamId/players/-temp-server-types';
+
+export const mockMembers: Member[] = [
+  {
+    id: 1,
+    name: '손흥민',
+    number: '#7',
+    position: 'FW',
+    foot: '왼발',
+    role: '일반',
+    joinDate: '2025-04-23',
+    profileImage: null,
+  },
+  {
+    id: 2,
+    name: '손흥민',
+    number: '#7',
+    position: 'MF',
+    foot: '오른발',
+    role: '일반',
+    joinDate: '2025-04-23',
+    profileImage: null,
+  },
+  {
+    id: 3,
+    name: '손흥민',
+    number: '#7',
+    position: 'DF',
+    foot: '양발',
+    role: '일반',
+    joinDate: '2025-04-23',
+    profileImage: null,
+  },
+  {
+    id: 4,
+    name: '손흥민',
+    number: '#7',
+    position: 'GK',
+    foot: '양발',
+    role: '일반',
+    joinDate: '2025-04-23',
+    profileImage: null,
+  },
+  {
+    id: 5,
+    name: '손흥민',
+    number: '#7',
+    position: 'FW',
+    foot: '양발',
+    role: '일반',
+    joinDate: '2025-04-23',
+    profileImage: null,
+  },
+  {
+    id: 6,
+    name: '손흥민',
+    number: '#7',
+    position: 'FW',
+    foot: '양발',
+    role: '일반',
+    joinDate: '2025-04-23',
+    profileImage: null,
+  },
+  {
+    id: 7,
+    name: '손흥민',
+    number: '#7',
+    position: 'FW',
+    foot: '양발',
+    role: '일반',
+    joinDate: '2025-04-23',
+    profileImage: null,
+  },
+  {
+    id: 8,
+    name: '손흥민',
+    number: '#7',
+    position: 'FW',
+    foot: '양발',
+    role: '일반',
+    joinDate: '2025-04-23',
+    profileImage: null,
+  },
+];
