+++ src/routes/teams/$teamId/players/-route.css.ts
@@ -6,9 +6,4 @@ export const container = style({
   backgroundColor: lightThemeVars.color.white.background,
   padding: '67px 52px 113px',
   overflow: 'hidden',
-  '@media': {
-    '(max-width: 991px)': {
-      padding: '67px 20px 100px',
-    },
-  },
 });
