+++ src/routes/teams/$teamId/players/route.tsx
@@ -5,6 +5,7 @@ import {
   MemberTable,
 } from '@/routes/teams/$teamId/players/-components';
 
+import { mockMembers } from './-mock-data';
 import * as styles from './-route.css';
 
 export const Route = createFileRoute('/teams/$teamId/players')({
@@ -15,7 +16,7 @@ function RouteComponent() {
   return (
     <div className={styles.container}>
       <MemberStatsSummary />
-      <MemberTable />
+      <MemberTable members={mockMembers} />
     </div>
   );
 }
